Apply package naming convention: discovery.azure -> cluster.discovery.azure
