+++ src/main/java/com/hivemq/extensions/cluster/discovery/azure/callback/AzureClusterDiscoveryCallback.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.discovery.azure.callback;
+package com.hivemq.extensions.cluster.discovery.azure.callback;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.LoggerContext;
@@ -25,9 +25,9 @@ import com.hivemq.extension.sdk.api.services.cluster.ClusterDiscoveryCallback;
 import com.hivemq.extension.sdk.api.services.cluster.parameter.ClusterDiscoveryInput;
 import com.hivemq.extension.sdk.api.services.cluster.parameter.ClusterDiscoveryOutput;
 import com.hivemq.extension.sdk.api.services.cluster.parameter.ClusterNodeAddress;
-import com.hivemq.extensions.discovery.azure.client.AzureStorageClient;
-import com.hivemq.extensions.discovery.azure.config.ClusterNodeFile;
-import com.hivemq.extensions.discovery.azure.config.ConfigReader;
+import com.hivemq.extensions.cluster.discovery.azure.client.AzureStorageClient;
+import com.hivemq.extensions.cluster.discovery.azure.config.ClusterNodeFile;
+import com.hivemq.extensions.cluster.discovery.azure.config.ConfigReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,7 +35,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import static com.hivemq.extensions.discovery.azure.util.StringUtil.isNullOrBlank;
+import static com.hivemq.extensions.cluster.discovery.azure.util.StringUtil.isNullOrBlank;
 
 /**
  * @author Till Seeberger
