+++ src/main/java/com/hivemq/extensions/cluster/discovery/azure/config/ConfigReader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.discovery.azure.config;
+package com.hivemq.extensions.cluster.discovery.azure.config;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
