+++ src/main/java/com/hivemq/extensions/cluster/discovery/azure/util/Preconditions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.discovery.azure.util;
+package com.hivemq.extensions.cluster.discovery.azure.util;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
