+++ src/main/java/com/hivemq/extensions/cluster/discovery/azure/util/StringUtil.java
@@ -13,20 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.discovery.azure;
 
-import org.jetbrains.annotations.NotNull;
-import org.testcontainers.utility.DockerImageName;
+package com.hivemq.extensions.cluster.discovery.azure.util;
 
-public final class DockerImageNames {
+import com.hivemq.extension.sdk.api.annotations.Nullable;
 
-    public static final @NotNull DockerImageName AZURITE_IMAGE = DockerImageName.parse("mcr.microsoft.com/azure-storage/azurite:3.14.3");
+/**
+ * @author Abdullah Imal
+ */
+public final class StringUtil {
 
-    public static final @NotNull DockerImageName TOXIPROXY_IMAGE = DockerImageName.parse("shopify/toxiproxy:2.1.0");
+    private StringUtil() {
+    }
 
-    public static final @NotNull DockerImageName HIVEMQ_IMAGE =
-            DockerImageName.parse("hivemq/hivemq4").withTag("latest");
-
-    public DockerImageNames() {
+    public static boolean isNullOrBlank(final @Nullable String value) {
+        return value == null || value.isBlank();
     }
 }
