+++ src/test/java/com/hivemq/extensions/cluster/discovery/azure/AzureClusterDiscoveryExtensionMainTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.discovery.azure;
+package com.hivemq.extensions.cluster.discovery.azure;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.parameter.*;
