+++ src/test/java/com/hivemq/extensions/cluster/discovery/azure/client/AzureStorageClientTest.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.hivemq.extensions.discovery.azure.client;
+package com.hivemq.extensions.cluster.discovery.azure.client;
 
 import com.azure.storage.blob.BlobClient;
 import com.azure.storage.blob.BlobContainerClient;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.parameter.ExtensionInformation;
-import com.hivemq.extensions.discovery.azure.config.ClusterNodeFileTest;
-import com.hivemq.extensions.discovery.azure.config.ConfigReader;
+import com.hivemq.extensions.cluster.discovery.azure.config.ClusterNodeFileTest;
+import com.hivemq.extensions.cluster.discovery.azure.config.ConfigReader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
